fix run scripts

+++ run.ps1
@@ -1,11 +1,19 @@
+param (
+    [string]$action = "start"
+)
+
 $ErrorActionPreference = "Stop"
 
 Copy-Item .env.example .env -Force
 
-$running = docker compose ps --status=running | Select-String "Up"
+if ($action -eq "stop") {
+    $running = docker compose ps --status=running | Select-String "Up"
 
-if ($running) {
-    docker compose down -v --remove-orphans
+    if ($running) {
+        docker compose down -v --remove-orphans
+    }
+
+    exit
 }
 
 docker compose build
