+++ run.sh
@@ -3,8 +3,13 @@ set -euo pipefail
 
 cp .env.example .env
 
-if docker compose ps --status=running | grep -q 'Up'; then
-    docker compose down -v --remove-orphans
+ACTION="${1:-start}"
+
+if [ "$ACTION" == "stop" ]; then
+    if docker compose ps --status=running | grep -q 'Up'; then
+        docker compose down -v --remove-orphans
+    fi
+    exit 0
 fi
 
 docker compose build
